+++ packages/core/babel-preset-dlight/src/pluginProvider.ts
@@ -1,5 +1,5 @@
 import type babel from "@babel/core"
-import { type types as t, type NodePath } from "@babel/core"
+import { types as t, type NodePath } from "@babel/core"
 import {
   type PropertyContainer,
   type HTMLTags,
@@ -15,6 +15,7 @@ import {
   devMode,
   dlightDefaultPackageName,
   importMap,
+  reactivityFuncNames,
 } from "./const"
 
 export class PluginProvider {
@@ -227,11 +228,10 @@ export class PluginProvider {
     //       @Watch(["count", "flag"])
     //       watcherFunc() { myFunc() }
     const watchDeco = this.findDecoratorByName(node.decorators, "Watch")
-    if (!watchDeco) {
-      if (this.t.isIdentifier(node.key, { name: "constructor" })) return
-      this.autoBindMethods(node)
-      return
-    }
+    if (this.t.isIdentifier(node.key, { name: "constructor" })) return
+    this.autoBindMethods(node)
+    if (!watchDeco) return
+
     // ---- Get dependencies from watcher decorator or watcher function decorator
     let deps: string[] = []
     if (this.t.isIdentifier(watchDeco)) {
@@ -394,15 +394,8 @@ export class PluginProvider {
   /**
    * @brief Decorator resolver: State
    * Add:
-   *  $${key} = ${value}
    *  $$${key} = ${depIdx}
    *  $sub$${key} = [${reversedDeps}]
-   *  get ${key}() {
-   *    return this.$${key}
-   *  }
-   *  set ${key}(value) {
-   *    this._$updateProp("${key}", value)
-   *  }
    * @param node
    */
   resolveStateDecorator(
@@ -413,7 +406,6 @@ export class PluginProvider {
     if (!this.classBodyNode) return
     if (!this.t.isIdentifier(node.key)) return
     const key = node.key.name
-    node.key.name = `$${key}`
     const propertyIdx = this.classBodyNode.body.indexOf(node)
 
     const idxNode = this.t.classProperty(
@@ -432,45 +424,7 @@ export class PluginProvider {
         ]
       : []
 
-    const getterNode = this.t.classMethod(
-      "get",
-      this.t.identifier(key),
-      [],
-      this.t.blockStatement([
-        this.t.returnStatement(
-          this.t.memberExpression(
-            this.t.thisExpression(),
-            this.t.identifier(`$${key}`)
-          )
-        ),
-      ])
-    )
-
-    const setterNode = this.t.classMethod(
-      "set",
-      this.t.identifier(key),
-      [this.t.identifier("value")],
-      this.t.blockStatement([
-        this.t.expressionStatement(
-          this.t.callExpression(
-            this.t.memberExpression(
-              this.t.thisExpression(),
-              this.t.identifier("_$updateProp")
-            ),
-            [this.t.stringLiteral(key), this.t.identifier("value")]
-          )
-        ),
-      ])
-    )
-
-    this.classBodyNode.body.splice(
-      propertyIdx + 1,
-      0,
-      idxNode,
-      ...depsNode,
-      getterNode,
-      setterNode
-    )
+    this.classBodyNode.body.splice(propertyIdx + 1, 0, idxNode, ...depsNode)
   }
 
   /* ---- Helper Functions ---- */
@@ -514,6 +468,7 @@ export class PluginProvider {
     const usedProperties = this.handleView()
     const propertyArr = Object.entries(this.propertiesContainer).reverse()
     const depReversedMap = this.dependencyMapReversed()
+    this.addAutoUpdate(usedProperties)
 
     for (const [
       key,
@@ -544,6 +499,169 @@ export class PluginProvider {
         )
       }
     }
+  }
+
+  private addAutoUpdate(usedProperties: string[]) {
+    if (!this.classBodyNode) return
+    const nonViewNodes = this.classBodyNode.body.filter(
+      n =>
+        !(
+          ((this.t.isClassProperty(n) || this.t.isClassMethod(n)) &&
+            ["View", "constructor", "_$compName"].includes(
+              (n.key as t.Identifier).name
+            )) ||
+          (this.t.isClassMethod(n, { kind: "method" }) &&
+            this.findDecoratorByName(n.decorators, "View"))
+        )
+    )
+    nonViewNodes.forEach(n => {
+      const value = this.t.isClassProperty(n)
+        ? n.value
+        : this.t.isClassMethod(n)
+          ? n.body
+          : null
+      if (!value) return
+      this.addAutoUpdateToNode(value, usedProperties)
+    })
+  }
+
+  private addAutoUpdateToNode(
+    node: t.Expression | t.BlockStatement,
+    usedProperties: string[]
+  ) {
+    this.traverse(this.valueWrapper(node), {
+      MemberExpression: path => {
+        if (
+          !this.t.isThisExpression(path.node.object) ||
+          !this.t.isIdentifier(path.node.property)
+        )
+          return
+        const key = path.node.property.name
+        if (!usedProperties.includes(key)) return
+        const assignPath = this.isAssignmentExpressionLeft(path)
+        if (!assignPath) return
+        this.addUpdate(assignPath, key)
+      },
+      CallExpression: path => {
+        if (!this.t.isMemberExpression(path.node.callee)) return
+        const funcNameNode = path.node.callee.property
+        if (!this.t.isIdentifier(funcNameNode)) return
+        if (!reactivityFuncNames.includes(funcNameNode.name)) return
+        let callee = path.get("callee").get("object") as NodePath
+
+        while (this.t.isMemberExpression(callee.node)) {
+          callee = callee.get("object") as NodePath
+        }
+        if (!this.t.isThisExpression(callee?.node)) return
+        const propName = (
+          (callee.parentPath!.node as t.MemberExpression)
+            .property as t.Identifier
+        ).name
+        this.addUpdate(path, propName)
+      },
+    })
+  }
+
+  private addUpdate(path: NodePath, key: string) {
+    // ---- this._$updateView("key")
+    const updateViewNode = this.t.callExpression(
+      this.t.memberExpression(
+        this.t.thisExpression(),
+        this.t.identifier("_$updateView")
+      ),
+      [this.t.stringLiteral(key)]
+    )
+    // ---- this._$updateDerived("key")
+    const updateDerivedNode = this.t.callExpression(
+      this.t.memberExpression(
+        this.t.thisExpression(),
+        this.t.identifier("_$updateDerived")
+      ),
+      [this.t.stringLiteral(key)]
+    )
+    // ---- Add updateDerived right on the assignment
+    let newNode: t.Node = this.t.sequenceExpression([
+      path.node as t.UpdateExpression | t.AssignmentExpression,
+      updateDerivedNode,
+    ])
+
+    // ---- Find the outermost block statement to add updateView, avoid
+    //      1. if block
+    //      2. loop block
+    let parentPath = path.parentPath
+    while (parentPath) {
+      if (
+        this.t.isBlockStatement(parentPath.node) &&
+        !this.t.isIfStatement(parentPath.parentPath?.node) &&
+        !this.t.isLoop(parentPath.parentPath?.node)
+      )
+        break
+      if (this.t.isFunction(parentPath.node)) {
+        parentPath = parentPath.get("body") as NodePath
+        break
+      }
+      parentPath = parentPath.parentPath as NodePath
+    }
+
+    // ---- Add updateView to last
+    if (this.t.isBlockStatement(parentPath?.node)) {
+      const node = parentPath.node
+      if (this.getUpdatePropExp(node).has(key)) return
+      const returns = this.getAllTopLevelReturnBlock(node)
+      returns.forEach(p => {
+        ;(p.node as t.BlockStatement).body.splice(
+          -1,
+          0,
+          this.t.expressionStatement(updateViewNode)
+        )
+      })
+      if (!this.t.isReturnStatement(node.body[node.body.length - 1])) {
+        node.body.push(this.t.expressionStatement(updateViewNode))
+      }
+    } else {
+      // ---- If no block statement found, do (original, update)
+      newNode = this.t.sequenceExpression([newNode, updateViewNode])
+    }
+
+    path.replaceWith(newNode)
+    path.skip()
+  }
+
+  private getUpdatePropExp(blockStatement: t.BlockStatement) {
+    const propNames = new Set()
+    blockStatement.body.forEach(node => {
+      if (!this.t.isExpressionStatement(node)) return
+      const exp = node.expression
+      if (!this.t.isCallExpression(exp)) return
+      const callee = exp.callee
+      if (!this.t.isMemberExpression(callee)) return
+      const obj = callee.object
+      const prop = callee.property
+      if (
+        !this.t.isThisExpression(obj) ||
+        !this.t.isIdentifier(prop, { name: "_$updateProp" })
+      )
+        return
+      const arg = exp.arguments[0]
+      if (!this.t.isStringLiteral(arg)) return
+      propNames.add(arg.value)
+    })
+    return propNames
+  }
+
+  private addViewAutoUpdate(obj: object) {
+    if (typeof obj !== "object") return
+    Object.entries(obj).forEach(([, value]) => {
+      if (!value) return
+      // ---- Type start with lowercase is not a node
+      if (value.type && value.type[0] !== value.type[0].toLowerCase()) {
+        if (this.t.isExpression(value)) {
+          this.addAutoUpdateToNode(value, this.availableProperties)
+        }
+      } else {
+        this.addViewAutoUpdate(value)
+      }
+    })
   }
 
   /* ---- DLight Class View Handlers ---- */
@@ -661,11 +779,13 @@ export class PluginProvider {
       subviewNames: subViewNames,
       htmlTags: this.htmlTags,
     })
+    viewUnits.map(viewUnit => this.addViewAutoUpdate(viewUnit))
 
     const [viewParticles, usedPropertySet] = parseReactivity(viewUnits, {
       babelApi: this.babelApi,
       availableProperties: this.availableProperties,
       dependencyMap: this.dependencyMap,
+      reactivityFuncNames,
     })
 
     const [body, classProperties, templateIdx] = generateView(viewParticles, {
@@ -710,12 +830,15 @@ export class PluginProvider {
       subviewNames: subViewNames,
       htmlTags: this.htmlTags,
     })
+    viewUnits.map(viewUnit => this.addViewAutoUpdate(viewUnit))
+
     const [viewParticlesProperty, usedPropertySet] = parseReactivity(
       viewUnits,
       {
         babelApi: this.babelApi,
         availableProperties: this.availableProperties,
         dependencyMap: this.dependencyMap,
+        reactivityFuncNames,
       }
     )
 
@@ -734,6 +857,7 @@ export class PluginProvider {
       dependencyMap: this.dependencyMap,
       dependencyParseType: "identifier",
       identifierDepMap,
+      reactivityFuncNames,
     })
 
     const subViewPropMap = Object.fromEntries(
@@ -848,7 +972,7 @@ export class PluginProvider {
    *  super()
    * }
    */
-  addConstructor(): t.ClassMethod {
+  private addConstructor(): t.ClassMethod {
     let constructor = this.classBodyNode!.body.find(n =>
       this.t.isClassMethod(n, { kind: "constructor" })
     ) as t.ClassMethod
@@ -867,7 +991,7 @@ export class PluginProvider {
     return constructor
   }
 
-  autoBindMethods(node: t.ClassMethod) {
+  private autoBindMethods(node: t.ClassMethod) {
     const constructorNode = this.addConstructor()
     constructorNode.body.body.push(
       this.t.expressionStatement(
@@ -892,7 +1016,7 @@ export class PluginProvider {
    *  return ${value}
    * }
    */
-  handleDerivedProperty(node: t.ClassProperty) {
+  private handleDerivedProperty(node: t.ClassProperty) {
     if (!this.t.isIdentifier(node.key)) return
     const key = node.key.name
     const value = node.value
@@ -912,7 +1036,7 @@ export class PluginProvider {
    * @param path
    * @returns dependencies
    */
-  getDependencies(node: t.ClassMethod | t.ClassProperty): string[] {
+  private getDependencies(node: t.ClassMethod | t.ClassProperty): string[] {
     if (!this.t.isIdentifier(node.key)) return []
 
     // ---- Deps: console.log(this.count)
@@ -924,25 +1048,25 @@ export class PluginProvider {
     )
     this.traverse(wrappedNode, {
       MemberExpression: innerPath => {
-        if (!this.t.isIdentifier(innerPath.node.property)) return
+        if (
+          !this.t.isIdentifier(innerPath.node.property) ||
+          !this.t.isThisExpression(innerPath.node.object)
+        )
+          return
+
         const propertyKey = innerPath.node.property.name
-        if (this.isAssignmentExpressionLeft(innerPath)) {
+        if (
+          this.isAssignmentExpressionLeft(innerPath) ||
+          this.isAssignmentFunction(innerPath)
+        ) {
           assignDeps.add(propertyKey)
         } else if (
           this.availableProperties.includes(propertyKey) &&
-          this.t.isThisExpression(innerPath.node.object) &&
           !this.isMemberInEscapeFunction(
             innerPath,
             this.classDeclarationNode!
           ) &&
-          !this.isMemberInManualFunction(
-            innerPath,
-            this.classDeclarationNode!
-          ) &&
-          !this.isAssignmentExpressionRight(
-            innerPath,
-            this.classDeclarationNode!
-          )
+          !this.isMemberInManualFunction(innerPath, this.classDeclarationNode!)
         ) {
           deps.add(propertyKey)
           this.dependencyMap[propertyKey]?.forEach(deps.add.bind(deps))
@@ -1036,6 +1160,34 @@ export class PluginProvider {
       node.value = this.t.identifier("undefined")
     }
     node.value = this.t.arrowFunctionExpression([], node.value)
+  }
+
+  /**
+   * @brief Get all top level return statements in a block statement,
+   *  ignore nested function returns
+   * @param node
+   * @returns
+   */
+  getAllTopLevelReturnBlock(node: t.BlockStatement): NodePath[] {
+    const returns: NodePath[] = []
+
+    let inNestedFunction = false
+    this.traverse(this.valueWrapper(node), {
+      Function: path => {
+        if (inNestedFunction) return
+        inNestedFunction = true
+        path.skip()
+      },
+      ReturnStatement: path => {
+        if (inNestedFunction) return
+        returns.push(path.parentPath)
+      },
+      exit: path => {
+        if (this.t.isFunction(path.node)) inNestedFunction = false
+      },
+    })
+
+    return returns
   }
 
   /**
@@ -1163,45 +1315,42 @@ export class PluginProvider {
   /**
    * @brief Check if it's the left side of an assignment expression, e.g. this.count = 1
    * @param innerPath
-   * @returns is left side of an assignment expression
+   * @returns assignment expression
    */
-  isAssignmentExpressionLeft(innerPath: NodePath): boolean {
-    const parentNode = innerPath.parentPath?.node
+  isAssignmentExpressionLeft(innerPath: NodePath): NodePath | null {
+    let parentPath = innerPath.parentPath
+    while (parentPath && !this.t.isStatement(parentPath.node)) {
+      if (this.t.isAssignmentExpression(parentPath.node)) {
+        if (parentPath.node.left === innerPath.node) return parentPath
+        const leftPath = parentPath.get("left") as NodePath
+        if (innerPath.isDescendant(leftPath)) return parentPath
+      } else if (this.t.isUpdateExpression(parentPath.node)) {
+        return parentPath
+      }
+      parentPath = parentPath.parentPath
+    }
 
-    return (
-      (this.t.isAssignmentExpression(parentNode) &&
-        parentNode.left === innerPath.node) ||
-      this.t.isUpdateExpression(parentNode)
-    )
+    return null
   }
 
   /**
-   * @brief Check if a member expression is the right side of an assignment expression
-   *   e.g. this.count = this.count + 1
+   * @brief Check if it's a reactivity function, e.g. arr.push
    * @param innerPath
-   * @returns is the right side of an assignment expression
+   * @returns
    */
-  isAssignmentExpressionRight(
-    innerPath: NodePath<t.MemberExpression>,
-    stopNode: t.Node
-  ): boolean {
-    const currNode = innerPath.node
+  isAssignmentFunction(innerPath: NodePath): boolean {
+    let parentPath = innerPath.parentPath
 
-    let isRightExp = false
-    let reversePath: NodePath<t.Node> | null = innerPath.parentPath
-    while (reversePath && reversePath.node !== stopNode) {
-      if (this.t.isAssignmentExpression(reversePath.node)) {
-        const leftNode = reversePath.node.left as t.MemberExpression
-        const typeEqual = currNode.type === leftNode.type
-        const identifierEqual =
-          (currNode.property as t.Identifier).name ===
-          (leftNode.property as t.Identifier).name
-        isRightExp = typeEqual && identifierEqual
-      }
-      reversePath = reversePath.parentPath
+    while (parentPath && this.t.isMemberExpression(parentPath.node)) {
+      parentPath = parentPath.parentPath
     }
-
-    return isRightExp
+    if (!parentPath) return false
+    return (
+      this.t.isCallExpression(parentPath.node) &&
+      this.t.isMemberExpression(parentPath.node.callee) &&
+      this.t.isIdentifier(parentPath.node.callee.property) &&
+      reactivityFuncNames.includes(parentPath.node.callee.property.name)
+    )
   }
 
   /**