+++ packages/core/transpiler/reactivity-parser/src/parser.ts
@@ -40,6 +40,7 @@ export class ReactivityParser {
   private readonly dependencyMap: Record<string, string[]>
   private readonly identifierDepMap: Record<string, string[]>
   private readonly dependencyParseType
+  private readonly reactivityFuncNames
 
   private readonly escapeNamings = ["escape", "$"]
   private static readonly customHTMLProps = [
@@ -73,6 +74,7 @@ export class ReactivityParser {
     this.dependencyMap = config.dependencyMap
     this.identifierDepMap = config.identifierDepMap ?? {}
     this.dependencyParseType = config.dependencyParseType ?? "property"
+    this.reactivityFuncNames = config.reactivityFuncNames ?? []
   }
 
   /**
@@ -589,19 +591,23 @@ export class ReactivityParser {
     node: t.Expression | t.Statement
   ): number[] {
     const deps = new Set<string>()
+    const assignDeps = new Set<string>()
 
     const wrappedNode = this.valueWrapper(node)
     this.traverse(wrappedNode, {
       Identifier: innerPath => {
         const identifier = innerPath.node
         const idName = identifier.name
+        if (!this.availableProperties.includes(idName)) return
         if (
-          this.availableProperties.includes(idName) &&
+          this.isAssignmentExpressionLeft(innerPath) ||
+          this.isAssignmentFunction(innerPath)
+        ) {
+          assignDeps.add(idName)
+        } else if (
           this.isStandAloneIdentifier(innerPath) &&
           !this.isMemberInEscapeFunction(innerPath) &&
-          !this.isMemberInManualFunction(innerPath) &&
-          !this.isAssignmentExpressionLeft(innerPath) &&
-          !this.isAssignmentIdentifierExpressionRight(innerPath)
+          !this.isMemberInManualFunction(innerPath)
         ) {
           deps.add(idName)
           this.dependencyMap[idName]?.forEach(deps.add.bind(deps))
@@ -609,6 +615,7 @@ export class ReactivityParser {
       },
     })
 
+    assignDeps.forEach(deps.delete.bind(deps))
     deps.forEach(this.usedProperties.add.bind(this.usedProperties))
     return [...deps].map(dep => this.availableProperties.indexOf(dep))
   }
@@ -620,25 +627,38 @@ export class ReactivityParser {
    *  3. the member expression is not in an escape function
    *  4. the member expression is not in a manual function
    *  5. the member expression is not the left side of an assignment expression, which is an assignment expression
-   *  6. the member expression is not the right side of an assignment expression, which is an update expression
+   *  6. the member is not a pure function declaration
    * @param node
    * @returns dependency index array
    */
   private getPropertyDependencies(node: t.Expression | t.Statement): number[] {
+    if (
+      this.t.isFunctionExpression(node) ||
+      this.t.isArrowFunctionExpression(node)
+    )
+      return []
+
     const deps = new Set<string>()
+    const assignDeps = new Set<string>()
 
     const wrappedNode = this.valueWrapper(node)
     this.traverse(wrappedNode, {
       MemberExpression: innerPath => {
-        if (!this.t.isIdentifier(innerPath.node.property)) return
+        if (
+          !this.t.isIdentifier(innerPath.node.property) ||
+          !this.t.isThisExpression(innerPath.node.object)
+        )
+          return
         const propertyKey = innerPath.node.property.name
         if (
+          this.isAssignmentExpressionLeft(innerPath) ||
+          this.isAssignmentFunction(innerPath)
+        ) {
+          assignDeps.add(propertyKey)
+        } else if (
           this.availableProperties.includes(propertyKey) &&
-          this.t.isThisExpression(innerPath.node.object) &&
           !this.isMemberInEscapeFunction(innerPath) &&
-          !this.isMemberInManualFunction(innerPath) &&
-          !this.isAssignmentExpressionLeft(innerPath) &&
-          !this.isAssignmentPropertyExpressionRight(innerPath)
+          !this.isMemberInManualFunction(innerPath)
         ) {
           deps.add(propertyKey)
           this.dependencyMap[propertyKey]?.forEach(deps.add.bind(deps))
@@ -646,6 +666,7 @@ export class ReactivityParser {
       },
     })
 
+    assignDeps.forEach(deps.delete.bind(deps))
     deps.forEach(this.usedProperties.add.bind(this.usedProperties))
     return [...deps].map(dep => this.availableProperties.indexOf(dep))
   }
@@ -884,67 +905,39 @@ export class ReactivityParser {
    * @returns is left side of an assignment expression
    */
   private isAssignmentExpressionLeft(innerPath: NodePath): boolean {
-    const parentNode = innerPath.parentPath?.node
+    let parentPath = innerPath.parentPath
+    while (parentPath && !this.t.isStatement(parentPath.node)) {
+      if (this.t.isAssignmentExpression(parentPath.node)) {
+        if (parentPath.node.left === innerPath.node) return true
+        const leftPath = parentPath.get("left") as NodePath
+        if (innerPath.isDescendant(leftPath)) return true
+      } else if (this.t.isUpdateExpression(parentPath.node)) {
+        return true
+      }
+      parentPath = parentPath.parentPath
+    }
 
+    return false
+  }
+
+  /**
+   * @brief Check if it's a reactivity function, e.g. arr.push
+   * @param innerPath
+   * @returns
+   */
+  private isAssignmentFunction(innerPath: NodePath): boolean {
+    let parentPath = innerPath.parentPath
+
+    while (parentPath && this.t.isMemberExpression(parentPath.node)) {
+      parentPath = parentPath.parentPath
+    }
+    if (!parentPath) return false
     return (
-      (this.t.isAssignmentExpression(parentNode) &&
-        parentNode.left === innerPath.node) ||
-      this.t.isUpdateExpression(parentNode)
+      this.t.isCallExpression(parentPath.node) &&
+      this.t.isMemberExpression(parentPath.node.callee) &&
+      this.t.isIdentifier(parentPath.node.callee.property) &&
+      this.reactivityFuncNames.includes(parentPath.node.callee.property.name)
     )
-  }
-
-  /**
-   * @brief Check if a member expression is the right side of an assignment expression
-   *   e.g. this.count = this.count + 1
-   * @param innerPath
-   * @returns is the right side of an assignment expression
-   */
-  private isAssignmentPropertyExpressionRight(
-    innerPath: NodePath<t.MemberExpression>
-  ): boolean {
-    const currNode = innerPath.node
-
-    let isRightExp = false
-    let reversePath: NodePath<t.Node> | null = innerPath.parentPath
-    while (reversePath) {
-      if (this.t.isAssignmentExpression(reversePath.node)) {
-        const leftNode = reversePath.node.left as t.MemberExpression
-        const typeEqual = currNode.type === leftNode.type
-        const identifierEqual =
-          (currNode.property as t.Identifier).name ===
-          (leftNode.property as t.Identifier).name
-        isRightExp = typeEqual && identifierEqual
-      }
-      reversePath = reversePath.parentPath
-    }
-
-    return isRightExp
-  }
-
-  /**
-   * @brief Check if an identifier is the right side of an assignment expression
-   *   e.g. count = count + 1
-   * @param innerPath
-   * @returns is the right side of an assignment expression
-   */
-  private isAssignmentIdentifierExpressionRight(
-    innerPath: NodePath<t.Identifier>
-  ): boolean {
-    const currNode = innerPath.node
-
-    let isRightExp = false
-    let reversePath: NodePath<t.Node> | null = innerPath.parentPath
-    while (reversePath) {
-      if (this.t.isAssignmentExpression(reversePath.node)) {
-        const leftNode = reversePath.node.left as t.Identifier
-        const typeEqual = currNode.type === leftNode.type
-        const identifierEqual = currNode.name === leftNode.name
-        isRightExp = typeEqual && identifierEqual
-      }
-      reversePath = reversePath.parentPath
-    }
-
-    return isRightExp
   }
 
   /**