+++ packages/core/transpiler/view-generator/src/HelperGenerators/BaseGenerator.ts
@@ -188,12 +188,7 @@ export default class BaseGenerator {
   }
 
   get updateParams() {
-    return [
-      this.t.identifier("$changed"),
-      this.t.identifier("$key"),
-      this.t.identifier("$prevValue"),
-      this.t.identifier("$newValue"),
-    ]
+    return [this.t.identifier("$changed"), this.t.identifier("$key")]
   }
 
   /**