+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,21 @@
 # babel-preset-dlight
 
+## 1.0.0-beta.39
+
+### Patch Changes
+
+- feat: new deep reactivity
+- Updated dependencies
+  - @dlightjs/reactivity-parser@1.0.0-beta.9
+  - @dlightjs/view-generator@1.0.0-beta.35
+
+## 1.0.0-alpha.38
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/view-generator@1.0.0-alpha.34
+
 ## 1.0.0-alpha.37
 
 ### Patch Changes