+++ packages/core/dlight/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @dlightjs/dlight
 
+## 1.0.0-beta.4
+
+### Patch Changes
+
+- build
+
+## 1.0.0-beta.3
+
+### Patch Changes
+
+- build
+
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- fix: prevalue in for
+
+## 1.0.0-beta.69
+
+### Patch Changes
+
+- feat: new deep reactivity
+
+## 1.0.0-alpha.68
+
+### Patch Changes
+
+- build
+
+## 1.0.0-alpha.67
+
+### Patch Changes
+
+- feat: obj not equal
+
 ## 1.0.0-alpha.66
 
 ### Patch Changes