+++ packages/core/transpiler/reactivity-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/reactivity-parser
 
+## 1.0.0-beta.9
+
+### Patch Changes
+
+- feat: new deep reactivity
+
 ## 1.0.0-alpha.8
 
 ### Patch Changes